Refactored the codebase; added Axis
I've made major changes to the code base. It was getting unruly.  As for
the problem we've been having, my solution is with Axis.  This class
stores a Path and a Record and works on them accordingly.  The Axis
assumes Path provides Position[] and Time[] while Record provides
Acceleration[] and Velocity[] since the latter two are to be calculated
by the solver.  That is, when solving, the solutions are stored to Axis.
However, ParameterSet (now ParameterSetConverter) needs to be creating
the Position[] and Time[] arrays, which is not currently doing.